feat: buffer_simulator: Simulate buffer generation in batches of N seconds
In this implementation the loaded data is splitted in temporal batches of N seconds (defined in the attribute batch_temporal_length).

The generation of the matrix batch is not computed in a constant time sequence (ex. every 5 seconds) as it would happend in a real-time scenario. (This happend because the loading of a single file would be higher than this defined value), but it is not necesary to fullfil this requirement in order to compute whatever we want in this batches.