feat: :train2: :arrows_clockwise: :trophy: Implement Batch and Chunks Buffer Managers
This commit allows to concat a N batches, that creates a 'chunk', and also manages the concatenation of M chunks that creates a 'capture'.

This management is done for each defined S sections. The 'captures' could be of only one chunk or M chunks.

Take into consideration that in case that a train is captured for a long time, this capture will be saved in the RAM memory, and for the moment, there is not a limit of how much chunks we can concatenate.