feat: :100: :rocket:  Implement real-time buffer manager and json file manager
In this implementation the cluster data is not being saving until the capture is finish. Intead, when a train is detected in a section, an status activation flag will be set to one and the data will be saved in memory ROM (as a file) and freed from RAM memory, "on the fly". This avoids the risk of colapsing the RAM memory space in case a train is being detected during a log period of time.