feat: :bug :fish:  Implement a new way of computing files maximun size
The client can define the files maximun sizes. The maximun sizes determines now the chunk's buffer-size (the numer of batches that will be saved in memory if there is a train capture). This value were not being computed correctly.

In this implementation it is assumed that every section's cluster's buffer-size could be likely different and dependant on the each section's spatial points (defined by client), the batch's temporal points (defined by nervision's config parameters) and the defined maximun size.

Take into account that the resultant file size will be aproximate to the defined maximun size. (It has been seen that sometimes could be slightly higher or lower (+- 0.1 Mb error))