validation: :dizzy: Raises an error in case the clients defines an start and end marging higher than the actual section's buffer's size
IMPORTANT:

The start and end margins defines the number of seconds that a capture will be saved before and after a train is captured. This gives some start and end margin's to the waterfall containing a train track.

In case the clients defines to high start and end margins. It could happen that this margins surpasses the computed buffer size. If this happens, the tool will NEVER be able to detect a train, and we TOTALLY DO NOT WANT THAT.

To solve this issue, we MUST warn the client in case uses a too high start or end margins.