feat: Compute the maximun file size for given maximun waiting time
The time a client's must wait until a capture is taken (a file is saved in memory) since a train is detected, in other words, the total buffer's time, limits the maximum file size that a clients should input.

Not having this into account and letting the client putting any file maximum value would led to having to wait "a lot" of time until the buffer is filled (and then saved in memory in case a train is captured), making the impresion that the tool does't correctly works, when it actually does.

So, for this reason,  an upper file size limit is a MUST.