refactor: Use better naming convection to distinguish between different file's limit sizes
We should distinguish between "file's size' and "file's size limit".

The first one,  determines the size that file's should have (it could be slightly higher or lower). This value is determined by client to each section.

The second one is also determined by client but in indirect way. The client will specify the seconds that is ready to wait to get all the buffer filled and thus, get a file ready to be read. From this "waiting-time" value, the maximun or limit size that a file should have is computed.

Is in the clien't hand to determine a "waiting-time" lower than the time between one train arrival to the next one, in case he want's to have different trains separated in different captures.