validate: config: Limit the maximun file size defined by the client
The client should be conscious that given a too high time limit, increases the RAM consumption a lot, increasing the risk of memory overload. Apart from that, should not take maximun times higher than the train's frecuency (the time between the pass of one time and other).

In any case, is better to consider that the client would be completly idiot!