validate: The upper limit of any section index must be limited
The developer will be responsible of giving this limit. The limit should never be higher than the expected batch space samples.

It has been defined a limit in case we want to limit even more the valid space the clients has available to define sections, due to the "fadding" fenomena that is more intense in the final sections of the fiber. This fadding difficults the trains capture a could led to false positives during the detection.