feat: Round start and end margings to upper value
This makes starting and ending margings to have at least one batch of margin.

Still, in some situations, could happend to not have an ending margin. If the clients wants and end margin, it should simply define one at least higher than the each batch time. (in production is at this moment defined with 4096 indexes with 0.001 seconds each, 4.096 seconds)