feat: :train2:  Implement new train detection mode
In the previous mode we were considering the number of adjacent samples that should marked as 'valid' to trigger a 'train-detection'.

In the new mode, we are considering the total percentage of samples that has been marked as 'valid'.

The diffence comes that this new mode do not take into account if the adjacent samples were valid or not and that is dependant of the defined spatial range.

This new mode, is not "new" at all, due to is compleatly tested in IPV project and demostrates its validity. That's way we implemented this methodology.

Presumably, both methodologies are compleately analog, and will work in the same way.

For the moment this "new" one has been chosen as default one.

It is faster to avoid "if" conditionals when making these computations, so changing from one mode to other will be done hard-coded, and not with a config variable.